+++ src/enums/calibrate.ts
@@ -1,5 +1,6 @@
 export enum Calibrate {
   initalMenu,
+  deviceConfig,
   audioTest1,
   audioTest2,
   process,
